Move generic toml handling to factors crate
